+++ src/components/students/studentDetail/cards/StudentDetailCard.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react'
 import { Github, Linkedin } from '../../../../assets/svg'
 import { Stud1 as ProfilePicture } from '../../../../assets/img'
-import { TTag } from '../../../../interfaces/interfaces'
+import { TTag } from '../../../../../types'
 import { useAppSelector } from '../../../../hooks/ReduxHooks'
 import LoadingSpiner from '../../../atoms/LoadingSpiner'
