+++ src/components/students/studentDetail/studentProfile/studentProfileCards/editStudentProfile/additionalInformation/DragAndDropLanguages.tsx
@@ -2,7 +2,7 @@ import { FC } from 'react'
 import {
     TLanguage,
     TLanguageLevel,
-} from '../../../../../../../interfaces/interfaces'
+} from '../../../../../../../../types'
 import { useDragAndDropLanguagesHook } from '../../../../../../../hooks/useDragAndDropLanguagesHook'
 
 export type TDragAndDropLanguagesProps = {
