+++ tsconfig.json
@@ -17,7 +17,7 @@
     "jsx": "react-jsx",
     "types": ["vitest/globals", "@testing-library/jest-dom"]
   },
-  "include": ["./src/**/*.ts", "./src/**/*.tsx", "vite.config.ts"],
+  "include": ["./src/**/*.ts", "./src/**/*.tsx", "vite.config.ts", "types.ts"],
   "exclude": ["node_modules", "dist", "build", "public", "scripts"],
   "references": [{ "path": "./tsconfig.node.json" }]
 }
